Put toghether createsidebar and directorypicker functions
